Add mapper

+++ src/main/java/br/com/forttiori/mongodb/service/StudentService.java
@@ -4,16 +4,14 @@ package br.com.forttiori.mongodb.service;
 import br.com.forttiori.mongodb.exceptions.EntityNotFoundException;
 import br.com.forttiori.mongodb.model.StudentRequest;
 import br.com.forttiori.mongodb.model.StudentResponse;
+import br.com.forttiori.mongodb.model.mapper.RequestMapper;
+import br.com.forttiori.mongodb.model.mapper.ResponseMapper;
 import br.com.forttiori.mongodb.persistence.entity.Students;
 import br.com.forttiori.mongodb.persistence.repository.StudentRepository;
 
-import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.annotation.Validated;
 
-import javax.validation.Valid;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -23,13 +21,13 @@ public class StudentService {
 
     private final StudentRepository studentRepository;
 
-    private final ModelMapper mapper;
 
-    public StudentService(StudentRepository studentRepository, ModelMapper mapper) {
+    public StudentService(StudentRepository studentRepository) {
         this.studentRepository = studentRepository;
-        this.mapper = mapper;
     }
 
+
+    // Método para retornar todos os estudantes ou retornar por idade
     public List<Students> find(Integer age){
         if(age == null) return this.studentRepository.findAll();
         return studentRepository.findAll().stream().filter(students -> students.getAge().equals(age)).collect(Collectors.toList());
@@ -38,48 +36,50 @@ public class StudentService {
 
 
 
-
+    // Método para retornar um estudante pelo id
     public StudentResponse getStudentsById(String id){
+
      var getById = studentRepository.findById(id).orElseThrow(() -> new EntityNotFoundException(" Id not found: " + id));
-     return createResponse(getById);
+
+     return ResponseMapper.createResponse(getById);
     }
 
+
+    // Método para criar um estudante
     public StudentResponse createStudent(StudentRequest studentRequest){
-            return createResponse(studentRepository.save(createRequest(studentRequest)));
-    }
+        Students students = RequestMapper.createEntity(studentRequest);
+        studentRepository.save(students);
 
-    // Coverte minha classe Estudante para uma "EstudanteDTO"
-    public StudentResponse createResponse(Students students){
-        return mapper.map(students, StudentResponse.class);
-    }
-    // Coverte EstudanteDTO para classe Estudante
-    public Students createRequest(StudentRequest request){
-        return mapper.map(request, Students.class);
+        return ResponseMapper.createResponse(students);
     }
 
 
-    public void deleteAll(@Validated List<String> id){
-        if(id == null) {
-            studentRepository.deleteAll();
-        }else {
-            for (String i: id) {
-                    studentRepository.findById(i).orElseThrow(() -> new EntityNotFoundException("Id not found: " + i));
-                studentRepository.deleteById(i);
-            }
-        };
+    // Método para deletar uma lista de ids ou caso não receber nenhum id deletar todo o banco;
+    public void deleteAll( List<String> id){
+
+        if(id.equals(null)){ studentRepository.deleteAll();}else{
+            studentRepository.deleteAllById(id);
+
+
+        }
+
+
     }
 
-    public StudentResponse update(String id, StudentRequest request){
-        Optional<Students> studentsOptional = Optional.ofNullable(studentRepository.findById(id).orElseThrow(() -> new EntityNotFoundException(" Id not found: " + id)));
+    // Método para atualizar um estudante
+    public StudentResponse updateStudent(String id, StudentRequest request){
+        Students entity;
+        entity = studentRepository.findById(id).orElseThrow(() -> new EntityNotFoundException(" Id not found: " + id));
 
-        var studentModel = studentsOptional.get();
-        studentModel.setName(request.getName());
-        studentModel.setSubjects(request.getSubjects());
-        studentModel.setAge(request.getAge());
-        studentModel.setEmail(request.getEmail());
-        studentModel.setStartDate(request.getStartDate());
-        studentRepository.save(studentModel);
-        return createResponse(studentModel);
+        entity.setName(request.getName());
+        entity.setSubjects(request.getSubjects());
+        entity.setAge(request.getAge());
+        entity.setEmail(request.getEmail());
+
+        studentRepository.save(entity);
+
+        return ResponseMapper.createResponse(entity);
+
     }
 
 }
